added test ouput
this is the result of on a 6128, load"copycat.bas", followed by a list#8.  This is what is sent to the PC, received by the python code and output to a file.  As you can see, it is the listing of the BASIC code that in itself actually outputs a CAT of the DISC in the floppy drive to the printer - so yes, running this BASIC code also sends this output to the PC if you run the program.  awesomeness, all round!